requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
aiogram==3.21.0
requests==2.32.4
python-gnupg==0.5.4
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
selenium==4.34.2
undetected-chromedriver==3.5.5
tweepy==4.16.0
telethon==1.34.0
python-dotenv==1.0.0 